Move all tests into a dedicated namespace and folder. Update composer locks.
