+++ tests/TelegramBot/Tests/Commands/CommandTestCase.php
@@ -8,9 +8,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Tests\Unit\Commands;
+namespace Longman\TelegramBot\Tests\Commands;
 
-use Tests\Unit\TestCase;
+use Longman\TelegramBot\Tests\TestCase;
 use Longman\TelegramBot\Telegram;
 
 /**
