+++ tests/TelegramBot/Tests/Commands/UserCommands/HelpCommandTest.php
@@ -8,10 +8,10 @@
  * file that was distributed with this source code.
  */
 
-namespace Tests\Unit\Commands\UserCommands;
+namespace Longman\TelegramBot\Tests\Commands\UserCommands;
 
-use Tests\Unit\Commands\CommandTestCase;
-use Tests\TestHelpers;
+use Longman\TelegramBot\Tests\Commands\CommandTestCase;
+use Longman\TelegramBot\Tests\TestHelpers;
 use Longman\TelegramBot\Commands\UserCommands\HelpCommand;
 
 /**
