+++ tests/TelegramBot/Tests/Entities/InlineKeyboardButtonTest.php
@@ -8,10 +8,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Tests\Unit;
+namespace Longman\TelegramBot\Tests;
 
 use Longman\TelegramBot\Entities\InlineKeyboardButton;
-use Longman\TelegramBot\Exception\TelegramException;
 
 /**
  * @package         TelegramTest
