+++ tests/TelegramBot/Tests/Entities/ReplyToMessageTest.php
@@ -8,10 +8,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Tests\Unit;
+namespace Longman\TelegramBot\Tests;
 
 use \Longman\TelegramBot\Entities\Update;
-use \Longman\TelegramBot\Entities\ReplyToMessage;
 
 /**
  * @package         TelegramTest
