+++ tests/TelegramBot/Tests/Entities/UpdateTest.php
@@ -8,7 +8,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Tests\Unit;
+namespace Longman\TelegramBot\Tests;
 
 use \Longman\TelegramBot\Entities\Update;
